+++ src/paypal-checkout/paypal-checkout.js
@@ -215,6 +215,11 @@ var REQUIRED_PARAMS_FOR_START_VAULT_INITIATED_CHECKOUT = [
  */
 function PayPalCheckout(options) {
   this._merchantAccountId = options.merchantAccountId;
+  // TODO remove this requirement for it to be opt in.
+  // This feature is not yet GA, so we're intentionally making
+  // it opt in and not publicly documenting it yet. Once it's
+  // GA, we can remove the requirement to opt in to it
+  this._autoSetDataUserIdToken = Boolean(options.autoSetDataUserIdToken);
 }
 
 PayPalCheckout.prototype._initialize = function (options) {
@@ -480,6 +485,8 @@ PayPalCheckout.prototype._createPaymentResource = function (options, config) {
   var self = this;
   var endpoint = 'paypal_hermes/' + constants.FLOW_ENDPOINTS[options.flow];
 
+  delete this.intentFromCreatePayment;
+
   config = config || {};
 
   if (options.offerCredit === true) {
@@ -491,6 +498,10 @@ PayPalCheckout.prototype._createPaymentResource = function (options, config) {
       endpoint: endpoint,
       method: 'post',
       data: self._formatPaymentResourceData(options, config)
+    }).then(function (data) {
+      self.intentFromCreatePayment = options.intent;
+
+      return data;
     });
   }).catch(function (err) {
     var status;
@@ -774,8 +785,8 @@ PayPalCheckout.prototype.tokenizePayment = function (tokenizeOptions) {
   var shouldVault = true;
   var payload;
   var options = {
-    flow: tokenizeOptions.billingToken ? 'vault' : 'checkout',
-    intent: tokenizeOptions.intent
+    flow: tokenizeOptions.billingToken && !tokenizeOptions.paymentID ? 'vault' : 'checkout',
+    intent: tokenizeOptions.intent || this.intentFromCreatePayment
   };
   var params = {
     // The paymentToken provided by the PayPal JS SDK is the EC Token
@@ -850,12 +861,15 @@ PayPalCheckout.prototype.getClientId = function () {
 /**
  * Resolves when the PayPal SDK has been succesfully loaded onto the page.
  * @public
- * @param {object} [options] A configuration object to modify the query params on the PayPal SDK. A subset of the parameters are listed below. For a full list of query params, see the [PayPal docs](https://developer.paypal.com/docs/checkout/reference/customize-sdk/?mark=query#query-parameters).
+ * @param {object} [options] A configuration object to modify the query params and data-attributes on the PayPal SDK. A subset of the parameters are listed below. For a full list of query params, see the [PayPal docs](https://developer.paypal.com/docs/checkout/reference/customize-sdk/?mark=query#query-parameters).
  * @param {string} [options.client-id] By default, this will be the client id associated with the authorization used to create the Braintree component. When used in conjunction with passing `authorization` when creating the PayPal Checkotu componet, you can speed up the loading of the PayPal SDK.
  * @param {string} [options.intent="authorize"] By default, the PayPal SDK defaults to an intent of `capture`. Since the default intent when calling {@link PayPalCheckout#createPayment|`createPayment`} is `authorize`, the PayPal SDK will be loaded with `intent=authorize`. If you wish to use a different intent when calling {@link PayPalCheckout#createPayment|`createPayment`}, make sure it matches here. If `sale` is used, it will be converted to `capture` for the PayPal SDK. If the `vault: true` param is used, no default intent will be passed.
  * @param {string} [options.currency="USD"] If a currency is passed in {@link PayPalCheckout#createPayment|`createPayment`}, it must match the currency passed here.
  * @param {boolean} [options.vault] Must be `true` when using `flow: vault` in {@link PayPalCheckout#createPayment|`createPayment`}.
  * @param {string} [options.components=buttons] By default, the Braintree SDK will only load the PayPal smart buttons component. If you would like to load just the [messages component](https://developer.paypal.com/docs/business/checkout/add-capabilities/credit-messaging/), pass `messages`. If you would like to load both, pass `buttons,messages`
+ * @param {object} [options.dataAttributes] The data attributes to apply to the script. Any data attribute can be passed. A subset of the parameters are listed below. For a full list of data attributes, see the [PayPal docs](https://developer.paypal.com/docs/checkout/reference/customize-sdk/#script-parameters).
+ * @param {string} [options.dataAttributes.data-client-token] The client token to use in the script. (usually not needed)
+ * @param {string} [options.dataAttributes.csp-nonce] See the [PayPal docs about content security nonces](https://developer.paypal.com/docs/checkout/reference/customize-sdk/#csp-nonce).
  * @param {callback} [callback] Called when the PayPal SDK has been loaded onto the page. The second argument is the PayPal Checkout instance. If no callback is provided, the promise resolves with the PayPal Checkout instance when the PayPal SDK has been loaded onto the page.
  * @returns {(Promise|void)} Returns a promise if no callback is provided.
  * @example <caption>Without options</caption>
@@ -884,6 +898,13 @@ PayPalCheckout.prototype.loadPayPalSDK = function (options) {
 
   this._paypalScript = document.createElement('script');
 
+  // NEXT_MAJOR_VERSION
+  // this options object should have 2 properties:
+  // * queryParams
+  // * dataAttributes
+  // should make organizing this better than squashing
+  // all the query params at the top level and extracting
+  // the data attributes
   options = assign({}, {
     components: 'buttons'
   }, options);
@@ -898,16 +919,22 @@ PayPalCheckout.prototype.loadPayPalSDK = function (options) {
     loadPromise.resolve();
   };
 
+  if (this._autoSetDataUserIdToken && this._configuration.authorizationType === 'CLIENT_TOKEN') {
+    this._paypalScript.setAttribute('data-user-id-token', this._configuration.authorizationFingerprint.split('?')[0]);
+  }
+
+  if (options.dataAttributes) {
+    Object.keys(options.dataAttributes).forEach(function (attribute) {
+      this._paypalScript.setAttribute('data-' + attribute, options.dataAttributes[attribute]);
+    }.bind(this));
+
+    delete options.dataAttributes;
+  }
+
   if (options['client-id']) {
     idPromise = Promise.resolve(options['client-id']);
   } else {
-    idPromise = this.getClientId().then(function (id) {
-      if (this._configuration.gatewayConfiguration.environment !== 'production') {
-        return 'sb';
-      }
-
-      return id;
-    }.bind(this));
+    idPromise = this.getClientId();
   }
 
   idPromise.then(function (id) {
@@ -997,16 +1024,17 @@ PayPalCheckout.prototype._formatTokenizeData = function (options, params) {
   var clientConfiguration = this._configuration;
   var gatewayConfiguration = clientConfiguration.gatewayConfiguration;
   var isTokenizationKey = clientConfiguration.authorizationType === 'TOKENIZATION_KEY';
+  var isVaultFlow = options.flow === 'vault';
   var data = {
     paypalAccount: {
       correlationId: params.billingToken || params.ecToken,
       options: {
-        validate: options.flow === 'vault' && !isTokenizationKey && options.vault
+        validate: isVaultFlow && !isTokenizationKey && options.vault
       }
     }
   };
 
-  if (params.billingToken) {
+  if (isVaultFlow) {
     data.paypalAccount.billingAgreementToken = params.billingToken;
   } else {
     data.paypalAccount.paymentToken = params.paymentId;